complete receipt page

Related work items: #64

+++ dotnet_ECommerce/dotnet_ECommerce/Models/Interfaces/IPayment.cs
@@ -5,6 +5,6 @@ namespace dotnet_ECommerce.Models.Interfaces
 {
     public interface IPayment
     {
-        bool Run(decimal total);
+        bool Run(double total);
     }
 }
